+++ src/about.ts
@@ -1,0 +1,2 @@
+
+export const githubLink = "https://github.com/Jerry-licious/spec";