+++ changelog.md
@@ -3,6 +3,7 @@
 - Added a limit on searches. 
 - Added support for TikZ figures.
 - Fixed an inconsistency in search bar styling.
+- Added self-promotion. 
 
 
 # v0.1.3, 2026-03-11