+++ changelog.md
@@ -4,6 +4,7 @@
 - Added support for TikZ figures.
 - Fixed an inconsistency in search bar styling.
 - Added self-promotion. 
+- Added version self-reporting.
 
 
 # v0.1.3, 2026-03-11