+++ changelog.md
@@ -1,4 +1,4 @@
-# ???
+# v0.1.4, 2026-03-12
 
 - Added a limit on searches. 
 - Added support for TikZ figures.